Add files via upload
A graph is a mathematical data structure defined by two non-empty sets: a set of nodes and a
set of edges. Graphs can be used to represent a network of cities, wherein the nodes indicate
the cities and the edges indicate the distances by train between any two cities. Dijkstra’s
algorithm is a greedy algorithm for finding the shortest path between two nodes in a graph. In
this project, Dijkstra’s algorithm has been implemented to compute the shortest rail route
between two cities.

The problem at hand is, given a set of cities and the distances between them by direct train, to find the shortest route to
be taken from a specified starting city to all the other cities.